pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ma-prompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepseek-ai==0.0.1",
    "pydantic==2.5.2",
    "pytest==7.4.3",
    "python-dotenv==1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ma_prompt*"]
namespaces = false
